+++ .travis.yml
@@ -12,11 +12,10 @@ before_install:
   - pip install -U pip
   - pip install -U setuptools
   - pip install -U wheel
-  - pip install flake8
 install:
   - pip install tox-travis .[test]
 script:
-  - flake8 niet
+  - tox
   - python setup.py install
   - niet tests/samples/sample.yaml project.meta.name
   - niet tests/samples/sample.yaml .project.meta.name
