+++ setup.cfg
@@ -30,8 +30,9 @@ test =
     fixtures
     tox
 devel=
+    fixtures
     pbr
-    flake8
+    tox
 
 [entry_points]
 console_scripts =
