+++ tox.ini
@@ -1,0 +1,29 @@
+[tox]
+envlist = py27,py34,py35,py36,pep8
+
+[testenv]
+usedevelop = True
+deps = .[test]
+       future
+       PyYAML
+       nose
+commands = nosetests
+
+[testenv:pep8]
+basepython = python3.6
+deps = flake8
+       flake8-import-order
+       flake8-blind-except
+       flake8-builtins
+       flake8-docstrings
+       flake8-rst-docstrings
+       flake8-logging-format
+commands = flake8
+
+[flake8]
+exclude = .tox,.eggs
+show-source = true
+ignore = D100,D101,D102,D103,D104
+
+[travis]
+python = 3.6: py36, pep8
